+++ py_sql_cleaner/__init__.py
@@ -1,3 +1,3 @@
 """Tooling for cleaning embedded SQL in Python code."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"